[lldb][Windows] Append access-violation detail to lldb-server stop description
